C1-StorageImp-TT

+++ storage/internal/domain/product.go
@@ -1,11 +1,11 @@
 package domain
 
 type Product struct {
-	ID     int    `db:"ID"`
-	Name   string `db:"name"`
-	Type   string `db:"type"`
-	Price  int    `db:"price"`
-	Count  int    `db:"count"`
-	Code   string `db:"code"`
-	Public int8   `db:"public"`
+	ID     int     `db:"ID"`
+	Name   string  `db:"name"`
+	Type   string  `db:"type"`
+	Price  float64 `db:"price"`
+	Count  int     `db:"count"`
+	Code   string  `db:"code"`
+	Public int8    `db:"public"`
 }
